requirements.txt
pandas>=2.1
openpyxl>=3.1
PyYAML>=6.0
watchdog>=3.0
pytest>=7.4
